CMD Decimation

+++ ual_ecar_vehicle_controller/firmware/vehicle_controller2pc-structs.h
@@ -316,7 +316,7 @@ struct TFrame_CONTROL_SIGNAL_payload_t
 	uint32_t timestamp_ms_tenth;
 	int16_t Steer_control_signal;
 	uint16_t Throttle_control_signal;
-	uint16_t Encoder_signal;
+	int32_t Encoder_signal;
 };
 struct TFrame_CONTROL_SIGNAL : public TBaseFrame<TFrame_CONTROL_SIGNAL_payload_t>
 {
